requirements/base.txt
Django==3.1.3
bpython==0.16
psycopg2==2.7.4
requests[socks]
celery==3.1.25
redis==2.10.6
djangorestframework==3.12.2
django-cors-headers==3.5.0
